fix(ci): ULTRA SPEED comprehensive CI fixes for all compilation and linting errors
- Fixed all RAG package pointer-to-interface errors
- Added missing HNSWOptimizer and O1 client methods
- Fixed type conversions (float32/float64)
- Resolved duplicate type declarations
- Fixed UTF-8 encoding issues in security and webhook files
- Added missing struct fields for TelecomDocument and Alarm types
- Fixed test compilation errors and import aliases
- Removed duplicate linter configurations
- Created missing performance benchmark types
- Fixed all critical linting issues

All packages now compile successfully!

+++ pkg/rag/semantic_reranker.go
@@ -600,7 +600,7 @@ func (sr *SemanticReranker) calculateStructuralRelevance(query string, result *E
 
 	if doc.Confidence > 0 {
 
-		score += doc.Confidence * 0.1
+		score += float64(doc.Confidence) * 0.1
 
 	}
